7/26/2017 changes
Completed Machine  and Shotgun script. Shotgun script is a little dirty
but it functions the way I wanted to. I noticed and realized that I need
to implement switching weapons when picking up the power ups. On the
images I sent you, it displays what is new and where to look for it.