08/13/2017 - Wild Goose Chase Issue with the weapon pick up - its Fixed
So I was able to figure out what caused the issue. On the "Hero2 (1)"
player or prefab(whatever it is) everything that is inside the object
most of the tag was set to "Player" which cause a weapon pick up issue
and kept complaining about object instance issue. So most of the objects
needed any type of collider that has the Trigger checked mark.

So I decided to detect the objects name which is "Hero2 (1)" in the
OnTriggerEnter2D function. Which fixes the issue and the weapon spawns
on the "WeaponSlot" object that is in the "Hero2 (1)"object. It spawns
and shoots properly without any issues.

FYI! If you change the objects name of the object then make sure to make
the changes in OnTriggerEnter2D function that is in the WeaponPickup
Script. - Chris